requirements.txt
checkov
azure-mgmt-resource
azure-mgmt-storage
azure-mgmt-network
azure-mgmt-compute
azure-identity
pydantic
openai
anthropic
python-dotenv
fpdf2
